#3 Add `types` package and definitions for nouns
This commit adds the `types` package, which will contain the definitions
for the different types of words (e.g. nouns, verbs, etc.). It also adds
the definitions for nouns, as well as the general `Word` struct, which
will be used for all words. Only the `typeDef` field will be different
for each word type.