#3 Add initial `saol` package with support for nouns
This commit adds the initial `saol` package with support for nouns as
defined in #3. The package is not yet complete, however, it is a good
starting point for further development. It makes use of the `colly` Go
package to scrape the SAOL website. It utilizes the `types`, `utils`
packages defined for this project.

A future objective would be to stop using the `found` global state
variable, which is used to filter out the words that are not of the
desired type. Generally speaking, it is good practise to avoid using
global state variables.